requirements.txt
streamlit
pandas
ollama
faiss-cpu 
transformers 
ollama
sentence_transformers